Removing the block styles class.
