+++ includes/classes/class-block-styles.php
@@ -1,0 +1,63 @@
+<?php
+namespace LSX\Classes;
+
+/**
+ * All the functions that run on the frontend and the rendering of the blocks.
+ *
+ * @package   LSX
+ * @author    LightSpeed
+ * @license   GPL3
+ * @link
+ * @copyright 2023 LightSpeed
+ */
+class Block_Styles {
+
+	/**
+	 * Contructor
+	 */
+	public function __construct() {
+	}
+
+	/**
+	 * Initiate our class.
+	 *
+	 * @return void
+	 */
+	public function init() {
+		add_action( 'init', array( $this, 'lsx_register_block_styles' ), 10 );
+	}
+
+	/**
+	 * Register our block styles.
+	 *
+	 * @return void
+	 */
+	public function register_block_styles() {
+
+		$block_styles = array(
+			'core/button'           => array(
+				'shadow'       => __( 'Shadow', 'lsx' ),
+			),
+		);
+
+		$block_styles = array(
+			'core/quote'           => array(
+				'underline-primary-secondary'  => __( 'Underline Primary-Secondary', 'lsx' ),
+				'underline-secondary-tertiary' => __( 'Underline Secondary-Tertiary', 'lsx' ),
+			),
+		);
+
+	
+		foreach ( $block_styles as $block => $styles ) {
+			foreach ( $styles as $style_name => $style_label ) {
+				register_block_style(
+					$block,
+					array(
+						'name'  => $style_name,
+						'label' => $style_label,
+					)
+				);
+			}
+		}
+	}
+}
