+++ includes/classes/class-core.php
@@ -4,6 +4,8 @@ namespace LSX;
 use LSX\Classes\Setup;
 use LSX\Classes\Block_Setup;
 use LSX\Classes\Block_Functions;
+use LSX\Classes\Block_Styles;
+use LSX\Classes\Frontend;
 
 /**
  * The main file loading the rest of the files
@@ -31,11 +33,25 @@ class Core {
 	public $block_setup;
 
 	/**
+	 * Contains the Block Styles.
+	 *
+	 * @var object
+	 */
+	public $block_styles;
+
+	/**
 	 * Contains the class responsible altering / displaying the blocks.
 	 *
 	 * @var object
 	 */
 	public $block_functions;
+
+	/**
+	 * All the functions for the output on the frontend.
+	 *
+	 * @var object
+	 */
+	public $frontend;
 
 	/**
 	 * Contructor
@@ -55,7 +71,9 @@ class Core {
 		// Initiate our classes.
 		$this->setup->init();
 		$this->block_setup->init();
+		$this->block_styles->init();
 		$this->block_functions->init();
+		$this->frontend->init();
 	}
 
 	/**
@@ -67,8 +85,14 @@ class Core {
 
 		require get_template_directory() . '/includes/classes/class-block-setup.php';
 		$this->block_setup = new Block_Setup();
+
+		require get_template_directory() . '/includes/classes/class-block-styles.php';
+		$this->block_styles = new Block_Styles();
 		
 		require get_template_directory() . '/includes/classes/class-block-functions.php';
 		$this->block_functions = new Block_Functions();
+
+		require get_template_directory() . '/includes/classes/class-frontend.php';
+		$this->frontend = new Frontend();
 	}
 }
